change filteredRestaurantsSpecification to use categories and dish names lists

+++ src/main/java/com/codecool/gastro/dto/review/NewReviewDto.java
@@ -13,7 +13,7 @@ public record NewReviewDto(
         String comment,
         @Min(value = 1, message = "Grade must be greater then or equal 1")
         @Max(value = 10, message = "Grade must be less then or equal 10")
-        Integer grade,
+        Double grade,
         @CustomerExist
         UUID customerId,
         @RestaurantExist
